Collapsed MongoCommandUtil into DefaultMongoCommandTagsProvider (same for its unit tests).
